PyPost-main/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.8.0
google-auth>=2.23.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
pandas>=2.0.0
requests>=2.31.0
urllib3>=2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2 